code cleaning to logical related concepts together and a similar restructuring for tests related to text
